App debugging with the FW 4.x memory layout
Use the g_(app|worker)_load_address variables which will exist in the
next FW release to locate the dynamic load addresses of the app and
worker when debugging an app under a FW 4.x+ emulator. The app and
worker need to be running when the debugger is attached in order to
retrieve the correct load addresses.

Tell GDB the addresses of all loaded sections, not just .text, .data and
.bss, so that it doesn't mistakenly think that e.g. .header is loaded at
address 0x0 anymore.